Backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
boto3==1.42.57
botocore==1.42.57
brotli==1.2.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==46.0.5
cssselect2==0.9.0
dj-database-url==3.1.2
Django==6.0.2
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-redis==6.0.0
django-storages==1.14.6
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fonttools==4.61.1
geographiclib==2.1
geopy==2.4.1
hiredis==3.3.0
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.6
zopfli==0.4.1
